mm: thp: cleanup and optimize compound_nr
Minor cleanup and minor optimizations to compound_nr and its callers.

There's no reason to return "unsigned long" in compound_nr, when it's
defined as "unsigned int" in the page structure and it's occasionally
truncated back to a 32bit value in its callers.

In addition its retval should be allowed to be cached in a caller
saved register and not be forced to be re-read from memory at every
iteration if any function called in the loop isn't in scope or inline.

+++ arch/arm64/kernel/mte.c
@@ -62,7 +62,7 @@ static void mte_sync_page_tags(struct page *page, pte_t old_pte,
 void mte_sync_tags(pte_t old_pte, pte_t pte)
 {
 	struct page *page = pte_page(pte);
-	long i, nr_pages = compound_nr(page);
+	unsigned int i, nr_pages = compound_nr(page);
 	bool check_swap = nr_pages == 1;
 	bool pte_is_tagged = pte_tagged(pte);
